fix: refactor registration component to use new Register component
