+++ __tests__/register.test.tsx
@@ -4,7 +4,7 @@ import { render, screen } from "@testing-library/react";
 import userEvent from '@testing-library/user-event'
 
 import QueryProviderWrapper from "@/app/QueryProviderWrapper";
-import RegisterPage from "@/app/register/page";
+import { Register } from '@/features/Login/Register/Register';
 import { ERROR_CREATE_USER_MESSAGE, ERROR_CREATE_USER_TITLE, SUCCESS_CREATE_USER_MESSAGE, SUCCESS_CREATE_USER_SPAN,
   SUCCESS_CREATE_USER_TITLE, SUCESS_CREATE_USER_SECONDARY_MESSAGE
  } from '@/shared/constants/Login.constants';
@@ -28,7 +28,7 @@ describe('Register', () => {
   it ('Show register page', () => {
     render(
       <QueryProviderWrapper>
-        <RegisterPage />
+        <Register />
       </QueryProviderWrapper>
     )
 
@@ -44,7 +44,7 @@ describe('Register', () => {
     const user = userEvent.setup()
     render(
       <QueryProviderWrapper>
-        <RegisterPage />
+        <Register />
       </QueryProviderWrapper>
     )
     const firstNameInput = screen.getByTestId('firstName')
@@ -92,7 +92,7 @@ describe('Register', () => {
 
       render(
         <QueryProviderWrapper>
-          <RegisterPage />
+          <Register />
         </QueryProviderWrapper>
       )
 
@@ -147,7 +147,7 @@ describe('Register', () => {
 
       render(
         <QueryProviderWrapper>
-          <RegisterPage />
+          <Register />
         </QueryProviderWrapper>
       )
 
@@ -193,7 +193,7 @@ describe('Register', () => {
 
       render(
         <QueryProviderWrapper>
-          <RegisterPage />
+          <Register />
         </QueryProviderWrapper>
       )
